Have chrome soldier on if /proc/cpuinfo lacks SSE2
When running under uml (for Travis), /proc/cpuinfo is all wrong and
the /opt/google/chrome/google-chrome script thinks that Chrome will be
unable to run, because the SSE2 instruction set seems to be missing on
the platform. Neuter this logic in the wrapper script, so that unit
tests may finally run on Travis (in Docker, itself in uml).

Rebuilt and re-pushed epflsti/blueboxnoc-tests